Initial implementation of volumetric photon maps. The idea is based on the Photon Beams described in "A Comprehensive Theory of Volumetric Radiance Estimation using Photon Points and Beams" by Jarosz 2011
This is a workable (but not completely finished) implementation of the volumetric photon maps. Although it does work well on some scenes (e.g. smoke-2.pbrt) but not in others (e.g. spotfog.pbrt). Particularly, the implemented approach of determining beam width doesn't work well in large open volumes and individual beams are clearly visible even with large number of photon paths. The approach is also very slow (perhaps, needs a better accelerator structure).

I'm keeping this implementation in a branch for now until I figure out what to do with it.